requirements.txt
Django==4.1.2
mysqlclient==2.1.1
psycopg2==2.9.5
numpy==1.23.5
pandas==1.5.2
matplotlib==3.6.2
django-bootstrap4==22.3
django-crispy-forms==1.14.0
#networkx python -> gephi output